backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
tzdata>=2024.1
pandas>=2.0.0
openpyxl>=3.1.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
python-multipart>=0.0.6
httpx>=0.27.0
